Leave Management System/requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
psycopg2-binary==2.9.7
gunicorn==21.2.0
python-dotenv==1.0.0
email-validator
flask-sqlalchemy